move adapters into the correct package
