+++ org.knime.scijava.commands/src/org/knime/scijava/commands/adapter/basic/CharOutputAdapter.java
@@ -1,7 +1,7 @@
 package org.knime.scijava.commands.adapter.basic;
 
 import org.knime.core.data.def.StringCell;
-import org.knime.scijava.commands.AbstractOutputAdapter;
+import org.knime.scijava.commands.adapter.AbstractOutputAdapter;
 import org.knime.scijava.commands.adapter.OutputAdapter;
 import org.scijava.plugin.Plugin;
